+++ tests/utils.cpp
@@ -581,4 +581,31 @@ TEST(utils, gfp_flags_format)
   EXPECT_EQ(GFPFlags::format(0x80000001), "GFP_DMA|0x80000000");
 }
 
+TEST(utils, path_ends_with)
+{
+  EXPECT_TRUE(path_ends_with("", ""));
+  EXPECT_TRUE(path_ends_with("/", "/"));
+  EXPECT_TRUE(path_ends_with("/a", "/a"));
+  EXPECT_TRUE(path_ends_with("/a/b/c", "c"));
+  EXPECT_TRUE(path_ends_with("/a/b/c", "b/c"));
+  EXPECT_TRUE(path_ends_with("/a/b/c", "a/b/c"));
+  EXPECT_TRUE(path_ends_with("/a/b/c", "/a/b/c"));
+  EXPECT_TRUE(path_ends_with("/lib64/libbpf.so", "libbpf.so"));
+  EXPECT_TRUE(path_ends_with(".", "."));
+  EXPECT_TRUE(path_ends_with("..", ".."));
+  EXPECT_TRUE(path_ends_with("/b/./c", "./c"));
+  EXPECT_TRUE(path_ends_with("/b/../c", "../c"));
+  EXPECT_TRUE(path_ends_with("./a/b/c", "./a/b/c"));
+  EXPECT_FALSE(path_ends_with("/", ""));
+  EXPECT_FALSE(path_ends_with("", "/a"));
+  EXPECT_FALSE(path_ends_with("/a/b/c", "x"));
+  EXPECT_FALSE(path_ends_with("/a/b/c", "x/c"));
+  EXPECT_FALSE(path_ends_with("/a/b/c", "x/b/c"));
+  EXPECT_FALSE(path_ends_with("/a/b/c", "/a/x/c"));
+  EXPECT_FALSE(path_ends_with("/lib64/libbpf.so", ""));
+  EXPECT_FALSE(path_ends_with("..", "."));
+  EXPECT_FALSE(path_ends_with(".", ".."));
+  EXPECT_FALSE(path_ends_with("/a", "/a/b/.."));
+}
+
 } // namespace bpftrace::test::utils